# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PrepAI – AI Placement Mentor API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /study-plans:
    get:
      operationId: listStudyPlans
      tags: [studyPlans]
      summary: List study plans for the current user
      responses:
        "200":
          description: List of study plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudyPlan"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      operationId: createStudyPlan
      tags: [studyPlans]
      summary: Generate a personalized study plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudyPlanInput"
      responses:
        "201":
          description: Generated study plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyPlan"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /study-plans/{id}:
    get:
      operationId: getStudyPlan
      tags: [studyPlans]
      summary: Get a study plan by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Study plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyPlan"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # Analytics
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # Dashboard
...
    StudyPlan:
      type: object
      required: [id, userId, targetCompany, daysLeft, weeks, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        targetCompany:
          type: string
        daysLeft:
          type: integer
        weeks:
          type: array
          items:
            $ref: "#/components/schemas/StudyWeek"
        createdAt:
          type: string
          format: date-time
    StudyWeek:
      type: object
      required: [week, topics, description]
      properties:
        week:
          type: integer
        topics:
          type: array
          items:
            type: string
        description:
          type: string
    StudyPlanInput:
      type: object
      required: [targetCompany, daysLeft]
      properties:
        targetCompany:
          type: string
          minLength: 1
        daysLeft:
          type: integer
          minimum: 1

    # Analytics
...